Set vars when we are filtering on user or category
This saves us some guesswork after that if we need
to know if the current export is filtered on a user
or on a category.

For example we need that when listing all the categories
and in case we are filtering on one category, we don't
need to show any others.